pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "config"]

[project]
name = "falcon-base"
version = "0.1.0"
description = "Falcon WSGI application skeleton"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "falcon>=4.0.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.8.0",
    "webtest>=3.0.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
